feat: Remove menu folder and change main path to "/explore"
