+++ src/constants.ts
@@ -5,7 +5,7 @@ export const PYROSCOPE_APP_ID = plugin.id;
 export const PLUGIN_BASE_URL = `/a/${PYROSCOPE_APP_ID}`;
 
 export enum ROUTES {
-  PROFILES_EXPLORER_VIEW = '/profiles-explorer',
-  ADHOC_VIEW = '/ad-hoc',
+  EXPLORE = '/explore',
+  ADHOC = '/ad-hoc',
   SETTINGS = '/settings',
 }
