+++ src/plugin.json
@@ -42,11 +42,11 @@
   "includes": [
     {
       "type": "page",
-      "name": "Explore profiles",
-      "path": "/a/%PLUGIN_ID%/profiles-explorer",
-      "role": "Viewer",
+      "name": "Profiles",
+      "path": "/a/%PLUGIN_ID%/explore",
+      "action": "datasources:explore",
       "addToNav": true,
-      "defaultNav": false
+      "defaultNav": true
     }
   ],
   "extensions": {
