+++ src/shared/domain/reportInteraction.ts
@@ -67,7 +67,7 @@ export type Interactions = {
   g_pyroscope_app_user_settings_clicked: {};
 };
 
-const PROFILES_EXPLORER_PAGE_NAME = ROUTES.PROFILES_EXPLORER_VIEW.slice(1);
+const PROFILES_EXPLORER_PAGE_NAME = ROUTES.EXPLORE.slice(1);
 
 function getCurrentPage(): string {
   const { pathname } = new URL(window.location.toString());
